Move ClassificationLink to classification_service.py
