+++ application/cms/dimension_service.py
@@ -3,8 +3,6 @@ from sqlalchemy import null
 from sqlalchemy.orm.exc import NoResultFound
 
 from application import db
-
-# from application.cms.dimension_classification_service import dimension_classification_service
 from application.cms.exceptions import (
     DimensionNotFoundException,
     DimensionAlreadyExists,
