+++ tests/test_ethnicity_classification_link_builder.py
@@ -1,5 +1,4 @@
-from application.cms.classification_service import ClassificationService
-from application.cms.classification_link import ClassificationLink
+from application.cms.classification_service import ClassificationLink, ClassificationService
 from application.data.standardisers.ethnicity_classification_finder_builder import (
     ethnicity_standardiser_from_data,
     ethnicity_classification_from_data,
